Add Forgotten Nature translations, manual and direct
This mod is difficult to translate because the old block names are not
namespaced and some do not even have names (tile.null) or have ambigious
names with vanilla, so they have to be directly mapped by a fixed
numeric ID (dependent on your specific configuration). But this covers a
large part of the terrain if you have Forgotten Nature in your world.